backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
pydantic
chromadb
numpy
